refactor: remove standalone compat bridge modules
