+++ crates/iam/src/store.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 pub mod object;
-mod storage_compat;
 
 use crate::cache::Cache;
 use crate::error::Result;
