+++ crates/obs/src/lib.rs
@@ -64,7 +64,6 @@ mod error;
 mod global;
 mod logging;
 pub mod metrics;
-mod storage_compat;
 mod telemetry;
 
 pub use cleaner::*;
