+++ crates/obs/src/metrics/scheduler.rs
@@ -23,7 +23,6 @@
 //! - Process disk I/O metrics
 //! - Host network I/O metrics
 
-use super::super::storage_compat::obs_bucket_monitor_available;
 use crate::metrics::collectors::{
     AuditTargetStats,
     BucketReplicationBandwidthStats,
@@ -83,6 +82,7 @@ use crate::metrics::stats_collector::{
     collect_scanner_metric_stats, collect_system_cpu_and_memory_stats_with,
 };
 use rustfs_audit::audit_target_metrics;
+use rustfs_ecstore::api::global as ecstore_global;
 use rustfs_notify::{notification_metrics_snapshot, notification_target_metrics};
 use rustfs_utils::get_env_opt_u64;
 use serde::Serialize;
@@ -599,7 +599,7 @@ pub fn init_metrics_runtime(token: CancellationToken) {
         loop {
             tokio::select! {
                 _ = interval.tick() => {
-                    let monitor_available = obs_bucket_monitor_available();
+                    let monitor_available = ecstore_global::get_global_bucket_monitor().is_some();
                     let stats = collect_bucket_replication_bandwidth_stats();
 
                     let current_live_keys = repl_bw_live_keys(&stats);
