+++ crates/s3select-api/src/lib.rs
@@ -19,7 +19,6 @@ use std::fmt::Display;
 pub mod object_store;
 pub mod query;
 pub mod server;
-mod storage_compat;
 
 #[cfg(test)]
 mod test;
